[hw05-mongo] rename /api/user -> /api/users
