+++ node-2023-11/hw04_les07-webServers/api-tests/framework/services/index.js
@@ -1,9 +1,9 @@
-const { UserApi } = require('./User.service');
+const { UsersApi } = require('./Users.service');
 const { ResetApi } = require('./Reset.service');
 const { CourseApi } = require('./Course.service');
 
 const apiProvider = () => ({
-    users: () => new UserApi(),
+    users: () => new UsersApi(),
     courses: () => new CourseApi(),
     reset: () => new ResetApi()
 });
