+++ node-2023-11/hw04_les07-webServers/app.js
@@ -8,7 +8,7 @@ var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
 var apiRouter = require('./routes/api');
 var apiCourseRouter = require('./routes/apiCourse');
-var apiUserRouter = require('./routes/apiUser');
+var apiUsersRouter = require('./routes/apiUsers');
 var apiLessonRouter = require('./routes/apiLesson');
 var loginRouter = require('./routes/login');
 var courseRouter = require('./routes/course');
@@ -66,7 +66,7 @@ app.use(
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
 app.use('/api/course', apiCourseRouter);
-app.use('/api/user', apiUserRouter);
+app.use('/api/users', apiUsersRouter);
 app.use('/api/lesson', apiLessonRouter);
 app.use('/api', apiRouter);
 app.use('/login', loginRouter);
